Update package versions (#34)
Summary:
#### Describe the solution you'd like
There is a vulnerability in one of the packages. Updating them to the latest

Also instead of just saying NodeJS greater that 8, doing specific versions that are updated without vulnerabilities.
"^8.10.0 || ^10.13.0 || >=11.10.1"

T47222529
Pull Request resolved: https://github.com/fbsamples/original-coast-clothing/pull/34

Test Plan:
Pushed the changes to my staging server. it seems to be working ok.
https://www.messenger.com/t/303005080360340

Reviewed By: Riandy

Differential Revision: D17107099

Pulled By: jorgeluiso

fbshipit-source-id: e271741489160727f69f879fec483845355785f1