Add copyright notices to EJS files (#70)
Summary:
There were files missing the FB copyright notices.

#### Is your pull request related to a problem? Please describe
The javascript style will include the header once in the client script, while the EJS style of <% /* %> will hide it — thus ensuring that it is displayed only once to someone inspecting the source.

#### Describe the solution you'd like
Ensure that each file has copyright notices, and that the main app script shipped to web clients have it declared exactly once at the the top

#### Describe alternatives you've considered
using only ejs-styled comments so as to not display the copyright notice to the webclient — but it _may_ be a legal requirement to have it shown at least once.

Pull Request resolved: https://github.com/fbsamples/original-coast-clothing/pull/70

Test Plan:
The internal validation tool should stop complaining about missing copyright notices

#### Checklist
- [{emoji:2714} ] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [{emoji:2714} ] I have updated the documentation accordingly.
- [{emoji:2714} ] I have read the **[CONTRIBUTING](https://github.com/fbsamples/original-coast-clothing/blob/master/CONTRIBUTING.md)** document.
- [n/a] I have added tests to cover my changes.
- [n/a] All new and existing tests passed.
- [{emoji:2714} ] The title of my pull request is a short description of the requested changes.
- [n/a] I have added the label **enhancement** to my pull request.

Reviewed By: rubycalling

Differential Revision: D25692479

Pulled By: Barbog

fbshipit-source-id: a7af7c39e6e915777433194b19d05e645c0ff8d9