Lead generation ad reported (#110)
Summary:
Add new flow to support reporting complete leads.

This new flow can be tested via this demo Click To Messenger Ad that supports complete lead reporting for ad ranking:
https://www.facebook.com/ads/experience/confirmation/?experience_id=603186808257936

Note that this is different from the Lead Generation in Messenger Product. That one transfers control to the app once the lead is complete while this new one is a CTM ad that supports leads as a reported signal.
https://developers.facebook.com/docs/messenger-platform/discovery/lead-generation-ads-in-messenger

# Other Changes
* This PR also updates package dependency versions mainly to keep the code current and address security vulnerabilities of the dependencies.
* Lint version and lint spec updates to help dev setup as well as code formatting
* `yarn lint --fix`
* Translations for new prompts in the CTM to Lead flow
* Addresses EJS lint issue
* Merge with the production deployed bot supporting https://m.me/542998526103632 to make sure this code matches what you can experience

Pull Request resolved: https://github.com/fbsamples/original-coast-clothing/pull/110

Reviewed By: gitLee47

Differential Revision: D40047919

Pulled By: jorgeluiso

fbshipit-source-id: bf01b979eb2e0bb72ecb977270210a71e3b5ebca