New Referral Webhooks on CTM Lead Generation Ads (#113)
Summary:
Implement the new Referal Webhooks available to Lead Generation with Click to Messenger Ads

As announced on Nov 9, 2022 new referral webhooks are sent on Click To Messenger, Lead Generation Ads
https://developers.facebook.com/docs/messenger-platform/discovery/lead-generation-ads-in-messenger

This change supports the bot response after a lead is complete. It implements the response that happend after this lead ad ends:
https://www.facebook.com/ads/experience/confirmation/?experience_id=2170497373124589

The demo ad is also linked in the documentation mentioned above as the Sample Lead Generation Ad.

Pull Request resolved: https://github.com/fbsamples/original-coast-clothing/pull/113

Reviewed By: gitLee47

Differential Revision: D41205127

Pulled By: jorgeluiso

fbshipit-source-id: 4c1f58c61a7c9a69bd9eced8b20d758fbe67a950